Reorganize scanning code.
