+++ internal/codeintel/db/dumps.go
@@ -2,6 +2,7 @@ package db
 
 import (
 	"context"
+	"database/sql"
 	"time"
 
 	"github.com/keegancsmith/sqlf"
@@ -22,6 +23,48 @@ type Dump struct {
 	FinishedAt        *time.Time `json:"finishedAt"`
 	RepositoryID      int        `json:"repositoryId"`
 	Indexer           string     `json:"indexer"`
+}
+
+// scanDumps scans a slice of dumps from the return value of `*dbImpl.query`.
+func scanDumps(rows *sql.Rows, queryErr error) (_ []Dump, err error) {
+	if queryErr != nil {
+		return nil, queryErr
+	}
+	defer func() { err = closeRows(rows, err) }()
+
+	var dumps []Dump
+	for rows.Next() {
+		var dump Dump
+		if err := rows.Scan(
+			&dump.ID,
+			&dump.Commit,
+			&dump.Root,
+			&dump.VisibleAtTip,
+			&dump.UploadedAt,
+			&dump.State,
+			&dump.FailureSummary,
+			&dump.FailureStacktrace,
+			&dump.StartedAt,
+			&dump.FinishedAt,
+			&dump.RepositoryID,
+			&dump.Indexer,
+		); err != nil {
+			return nil, err
+		}
+
+		dumps = append(dumps, dump)
+	}
+
+	return dumps, nil
+}
+
+// scanFirstDump scans a slice of dumps from the return value of `*dbImpl.query` and returns the first.
+func scanFirstDump(rows *sql.Rows, err error) (Dump, bool, error) {
+	dumps, err := scanDumps(rows, err)
+	if err != nil || len(dumps) == 0 {
+		return Dump{}, false, err
+	}
+	return dumps[0], true, nil
 }
 
 // GetDumpIDs returns all dump ids in chronological order.
