+++ internal/codeintel/db/indexes.go
@@ -2,6 +2,7 @@ package db
 
 import (
 	"context"
+	"database/sql"
 	"time"
 
 	"github.com/keegancsmith/sqlf"
@@ -20,6 +21,51 @@ type Index struct {
 	FinishedAt        *time.Time `json:"finishedAt"`
 	RepositoryID      int        `json:"repositoryId"`
 	Rank              *int       `json:"placeInQueue"`
+}
+
+// scanIndexes scans a slice of indexes from the return value of `*dbImpl.query`.
+func scanIndexes(rows *sql.Rows, queryErr error) (_ []Index, err error) {
+	if queryErr != nil {
+		return nil, queryErr
+	}
+	defer func() { err = closeRows(rows, err) }()
+
+	var indexes []Index
+	for rows.Next() {
+		var index Index
+		if err := rows.Scan(
+			&index.ID,
+			&index.Commit,
+			&index.QueuedAt,
+			&index.State,
+			&index.FailureSummary,
+			&index.FailureStacktrace,
+			&index.StartedAt,
+			&index.FinishedAt,
+			&index.RepositoryID,
+			&index.Rank,
+		); err != nil {
+			return nil, err
+		}
+
+		indexes = append(indexes, index)
+	}
+
+	return indexes, nil
+}
+
+// scanFirstIndex scans a slice of indexes from the return value of `*dbImpl.query` and returns the first.
+func scanFirstIndex(rows *sql.Rows, err error) (Index, bool, error) {
+	indexes, err := scanIndexes(rows, err)
+	if err != nil || len(indexes) == 0 {
+		return Index{}, false, err
+	}
+	return indexes[0], true, nil
+}
+
+// scanFirstIndexInterface scans a slice of indexes from the return value of `*dbImpl.query` and returns the first.
+func scanFirstIndexInterface(rows *sql.Rows, err error) (interface{}, bool, error) {
+	return scanFirstIndex(rows, err)
 }
 
 // GetIndexByID returns an index by its identifier and boolean flag indicating its existence.
@@ -123,7 +169,7 @@ var indexColumnsWithNullRank = []*sqlf.Query{
 // If there is no such unlocked index, a zero-value index and nil DB will be returned along with a false
 // valued flag. This method must not be called from within a transaction.
 func (db *dbImpl) DequeueIndex(ctx context.Context) (Index, DB, bool, error) {
-	index, tx, ok, err := db.dequeueRecord(ctx, "lsif_indexes", indexColumnsWithNullRank, sqlf.Sprintf("queued_at"), scanFirstIndexDequeue)
+	index, tx, ok, err := db.dequeueRecord(ctx, "lsif_indexes", indexColumnsWithNullRank, sqlf.Sprintf("queued_at"), scanFirstIndexInterface)
 	if err != nil || !ok {
 		return Index{}, tx, ok, err
 	}
